Refactor shared thread query validation
