+++ apps/web/hooks/useThreads.ts
@@ -1,7 +1,7 @@
 import useSWR from "swr";
 import type { ThreadsResponse } from "@/app/api/threads/route";
 import type { Thread as EmailThread } from "@/components/email-list/types";
-import type { ThreadsQuery } from "@/app/api/threads/validation";
+import type { ThreadsQuery } from "@/utils/threads/validation";
 
 export type Thread = EmailThread;
